requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
cffi==1.15.0
cryptography==36.0.1
Deprecated==1.2.13
Django==3.2.12
django-cors-headers==3.11.0
django-filter==21.1
django-redis==5.2.0
django-simpleui==2022.3.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
loguru==0.6.0
packaging==21.3
pycparser==2.21
PyJWT==2.3.0
PyMySQL==1.0.2
pyparsing==3.0.7
pytz==2021.3
redis==4.1.4
sqlparse==0.4.2
wrapt==1.13.3
